perf(vcf/vca): preallocate buffers and use float exp2/powf where appropriate

+++ Source/CS01Synth/VCAProcessor.cpp
@@ -113,7 +113,7 @@ void VCAProcessor::processBlock(juce::AudioBuffer<float>& buffer, juce::MidiBuff
 // IG02600 VCA chip emulation
 float VCAProcessor::processVCA(float input, float controlVoltage, float volumeParam) {
     // Apply logarithmic volume control characteristic (PVR5)
-    float volume = std::pow(volumeParam, 2.5f);  // Approximate log curve
+    float volume = std::powf(volumeParam, 2.5f);  // Use float version for performance
 
     // Calculate gain based on control voltage and volume
     float gain = controlVoltage * volume;
